Moved support classes into org.grails.cache.redis.internal
