+++ src/main/groovy/grails/plugin/cache/redis/internal/GrailsDeserializingConverter.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package grails.plugin.cache.web.filter.redis;
+package grails.plugin.cache.redis.internal;
 
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.core.serializer.Deserializer;
